+++ apps/mobile/src/features/signIn/components/signIn/discord.tsx
@@ -1,39 +1,15 @@
-import { openAuthSessionAsync } from "expo-web-browser";
-import { Button } from "react-native";
-import { supabase } from "@/utils/supabase";
+import DiscordIcon from "@shokupass/ui/assets/social/discord.svg";
+import { Text, TouchableOpacity } from "react-native";
+import { signInWithDiscord } from "../../utils/signIn";
 
-const signInWithDiscord = async () => {
-  const { error, data } = await supabase.auth.signInWithOAuth({
-    provider: "discord",
-    options: {
-      redirectTo: "shokupass://discord-auth",
-    },
-  });
-  if (error) {
-    return;
-  }
-  const result = await openAuthSessionAsync(data.url, "shokupass://discord-auth?", {
-    showInRecents: true,
-  });
-
-  if (result && result.type === "success") {
-    const parsedUrl = new URL(result.url);
-    const urlParams = new URLSearchParams(parsedUrl.hash.replace("#", "?"));
-
-    const paramObject: Record<string, string> = {};
-    urlParams.forEach((value, key) => {
-      paramObject[key] = value;
-    });
-
-    if (paramObject["access_token"] && paramObject["refresh_token"]) {
-      await supabase.auth.setSession({
-        access_token: paramObject["access_token"],
-        refresh_token: paramObject["refresh_token"],
-      });
-    }
-  }
-};
-
-const DiscordSignInButton = () => <Button onPress={signInWithDiscord} title="Sign In with Discord" />;
+const DiscordSignInButton = () => (
+  <TouchableOpacity
+    className="box-border h-12 w-64  flex-row items-center gap-x-2 rounded-lg border-2 border-gray-300 p-0"
+    onPress={signInWithDiscord}
+  >
+    <DiscordIcon width={32} height={32} />
+    <Text className="flex-1 font-mPlusRounded1c text-xl">Sign in with Discord</Text>
+  </TouchableOpacity>
+);
 
 export { DiscordSignInButton };
