+++ apps/mobile/src/features/signIn/components/signIn/google.tsx
@@ -1,4 +1,6 @@
-import { GoogleSignin, GoogleSigninButton } from "@react-native-google-signin/google-signin";
+import { GoogleSignin } from "@react-native-google-signin/google-signin";
+import GoogleIcon from "@shokupass/ui/assets/social/google.svg";
+import { TouchableOpacity, Text } from "react-native";
 import { supabase } from "@/utils/supabase";
 
 const GoogleSignInButton = () => {
@@ -6,9 +8,8 @@ const GoogleSignInButton = () => {
     webClientId: process.env["EXPO_PUBLIC_GOOGLE_WEB_CLIENT_ID"]!,
   });
   return (
-    <GoogleSigninButton
-      size={GoogleSigninButton.Size.Standard}
-      color={GoogleSigninButton.Color.Dark}
+    <TouchableOpacity
+      className="box-border h-12 w-64 flex-row  items-center gap-x-2 rounded-lg border-2 border-gray-300 p-0"
       onPress={async () => {
         try {
           await GoogleSignin.hasPlayServices();
@@ -23,7 +24,10 @@ const GoogleSignInButton = () => {
           console.log("catch error", JSON.stringify(error, null, "\t"));
         }
       }}
-    />
+    >
+      <GoogleIcon width={32} height={32} />
+      <Text className="flex-1 font-mPlusRounded1c text-xl">Sign in with Google</Text>
+    </TouchableOpacity>
   );
 };
