+++ apps/mobile/src/features/signIn/utils/index.ts
@@ -1,2 +1,2 @@
-export { sighIn } from "./signIn";
-export { signOut } from "./signOut";
+export * from "./signIn";
+export * from "./signOut";
